Proposed README
Would you guys kindly consider filling up the `XXX` and accepting this PR? Since this arguably the most important experiment in this repo, I believe it deservers a little extra information. What do you think?

I'll try to rerun the experiments and fill it up myself if you are interested. A TODO for this PR should also be a CLI to generate sample translations.